Merge pull request #2749 from cloudflare/release-5.8.0

+++ src/cloudflare/types/radar/attacks/layer7/top/location_origin_response.py
@@ -126,7 +126,7 @@ class Top0(BaseModel):
 
     origin_country_name: str = FieldInfo(alias="originCountryName")
 
-    rank: float
+    rank: int
 
     value: str
